split up into one file per config to argument conversion
