+++ dockercmd/arguments/helper.go
@@ -1,0 +1,17 @@
+package arguments
+
+import (
+	"fmt"
+	"github.com/drone/envsubst"
+	"os"
+)
+
+func resolveEnvVar(envVarName string) (string, error) {
+	return envsubst.Eval(envVarName, func(normalizedEnvVarName string) string {
+		if envVar, ok := os.LookupEnv(normalizedEnvVarName); !ok {
+			panic(fmt.Sprintf("env var %v is not set!", normalizedEnvVarName))
+		} else {
+			return envVar
+		}
+	})
+}
