+++ dockercmd/arguments/impersonation_windows.go
@@ -5,13 +5,6 @@ import (
 	"github.com/Oppodelldog/droxy/dockercmd/builder"
 )
 
-func addGroups(commandDef *config.CommandDefinition, builder *builder.Builder) error {
-	_ = commandDef
-	_ = builder
-
-	return nil
-}
-
 func addImpersonation(commandDef *config.CommandDefinition, builder *builder.Builder) error {
 	_ = commandDef
 	_ = builder
