+++ dockercmd/arguments/specific_linux.go
@@ -39,29 +39,3 @@ func buildGroups(areGroupsAdded bool, builder *builder.Builder) error {
 
 	return nil
 }
-
-func addImpersonation(commandDef *config.CommandDefinition, builder *builder.Builder) error {
-	if impersonate, ok := commandDef.GetImpersonate(); ok {
-		err := buildImpersonation(impersonate, builder)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func buildImpersonation(isImpersonated bool, builder *builder.Builder) error {
-	if !isImpersonated {
-		return nil
-	}
-
-	currentUser, err := user.Current()
-	if err != nil {
-		return err
-	}
-
-	builder.SetContainerUserAndGroup(currentUser.Uid, currentUser.Gid)
-
-	return nil
-}
